Update GOL.java
Fixed the problem I had with checking each cell for its amt of neighbors where I would reveive an index out of bounds error. Also made a chenge so instead of the state of the cells being represented by a character, it is now repesented by a color. For now, the screen is updated every 4 seconds but I want to add a pause and start function to the window, along with a setting to set the speed at which the screen updates